Rename guid_support to winapi_support
